Endret en kommentar

+++ opg3.py
@@ -1,11 +1,11 @@
 # Tar inn tre tall fra brukeren
 a = abs(int(input("Tall1: ")))
-b = abs(int(input("Tall2: ")))
-c = abs(int(input("Tall3: ")))
+b = abs(float(input("Tall2: ")))
+c = abs(float(input("Tall3: ")))
 
 
 def triple(a, b, c):
-    # Endrer variablene til å være opphøyd i andre
+    # Endrer variablene til kvadratet av seg selv
     a = a**2
     b = b**2
     c = c**2
